Reconfigure menu to go to top on small displays.
The menu was originally configured to disappear when the screen was
small so you had to toggle the menu with the button. This changes the
behavior to move the menu to the top of the page for small displays.
In this case, the menu is also compressed to take up less space. I
like it much better.